requirement.txt
# Core framework
fastapi==0.111.0
uvicorn==0.30.1

# NLP and AI models
transformers==4.42.4
torch==2.3.0
sentencepiece==0.2.0
nltk==3.9.1
scikit-learn==1.5.0

# Speech recognition (for audio-to-text)
whisper==1.1.10
SpeechRecognition==3.10.0
pydub==0.25.1

# Summarization, QA, and RAG
faiss-cpu==1.8.0
langchain==0.2.0
chromadb==0.5.0

# Translation
googletrans==4.0.0-rc1
deep-translator==1.11.4

# Data handling
pandas==2.2.2
numpy==1.26.4

# Authentication + Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.3

# Database (SQLite for MVP)
sqlalchemy==2.0.31
databases==0.9.0

# Utilities
python-multipart==0.0.9
requests==2.32.3
beautifulsoup4==4.12.3
